Separate GoApp and GoFFApp
