+++ config/apps.master.json
@@ -157,7 +157,7 @@
       "url_prefix": ""
     },
     {
-      "app_type": "GoApp",
+      "app_type": "GoFFApp",
       "service": "FLOW_FRAMEWORK",
       "display_name": "Flow Framework",
       "repo_name": "flow-framework",
