+++ src/system/apps/base/__init__.py
@@ -1,6 +1,7 @@
-__all__ = ['frontend_app', 'python_app', 'go_app', 'apt_app']
+__all__ = ['frontend_app', 'python_app', 'go_ff_app', 'go_app', 'apt_app']
 
+from src.system.apps.base import apt_app
 from src.system.apps.base import frontend_app
 from src.system.apps.base import go_app
+from src.system.apps.base import go_ff_app
 from src.system.apps.base import python_app
-from src.system.apps.base import apt_app
