+++ src/system/apps/enums/enums.py
@@ -5,6 +5,7 @@ class Types(enum.Enum):
     INSTALLER = 'Installer'
     FRONTEND_APP = 'FrontendApp'
     PYTHON_APP = 'PythonApp'
+    GO_FF_APP = 'GoFFApp'
     GO_APP = 'GoApp'
     JAVA_APP = 'JavaApp'
     APT_APP = 'AptApp'
